api/requirements.txt
kafka-python
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
logging
python-dotenv
fastapi
uvicorn
prometheus-client